feat(game): add first build polygongame with default values 2 player 4 walls (classic pong)
in this build everything is already created for later. all states are saved in Redis. i have decided to save the data in Redis in the MessagePack “format”. depending on which game is selected, a different game_logic is applied. the backend is currently sending three different messages (initial_state, game_state, game_finished, error).
the values from the frontend for the game setup arrive in the backend but are not yet applied.